requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.1
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-tailwind==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
icecream==2.1.3
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
openai==1.10.0
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
